+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## main
 
+### Changed
+
+- `ZoneRecord.type`, `TemplateRecord.type`, and `createZoneRecord` type parameter changed from enum to `string`
+
 ### Removed
 
 - Removed deprecated `getDomainPremiumPrice`. Use `getDomainPrices` instead. (dnsimple/dnsimple-developer#916)
+- Removed `ZoneRecordType` and `TemplateRecordType` type enumerations
 - Removed deprecated `getWhoisPrivacy` (dnsimple/dnsimple-developer#919)
 - Removed deprecated `renewWhoisPrivacy` (dnsimple/dnsimple-developer#919)